requirements.txt
Flask~=3.1.1
requests~=2.32.4
python-docx
pandas~=2.3.1
openpyxl
pdfplumber~=0.11.7
dotenv~=0.9.9
python-dotenv~=1.1.1